chore: move spelling correction files to different directories
