+++ deeppavlov/configs/error_model/brillmoore_kartaslov_ru.json
@@ -41,7 +41,7 @@
         "load_path": "error_model/error_model_ru.tsv"
       },
       {
-        "class": "deeppavlov.models.spellers.error_model.kenlm_elector:KenlmElector",
+        "class": "deeppavlov.models.spelling_correction.electors.kenlm_elector:KenlmElector",
         "in": ["tokens_candidates"],
         "out": ["y_predicted_tokens"],
         "load_path": "language_models/ru_wiyalen_no_punkt.arpa.binary"
