+++ deeppavlov/configs/error_model/levenstein_searcher.json
@@ -23,11 +23,11 @@
       {
         "in": ["x_tokens"],
         "out": ["tokens_candidates"],
-        "class": "deeppavlov.models.spellers.error_model.leve_searcher:LeveSearcher",
+        "class": "deeppavlov.models.spelling_correction.levenstein.searcher_component:LevensteinSearcherComponent",
         "words": "#vocab.keys()"
       },
       {
-        "class": "deeppavlov.models.spellers.error_model.kenlm_elector:KenlmElector",
+        "class": "deeppavlov.models.spelling_correction.electors.kenlm_elector:KenlmElector",
         "in": ["tokens_candidates"],
         "out": ["y_predicted_tokens"],
         "load_path": "language_models/ru_wiyalen_no_punkt.arpa.binary"
