+++ deeppavlov/models/spelling_correction/levenstein/searcher_component.py
@@ -20,21 +20,21 @@ from deeppavlov.core.common.registry import register
 from deeppavlov.core.models.component import Component
 from deeppavlov.core.common.log import get_logger
 
-from .searcher.levenstein_searcher import LevensteinSearcher
+from .levenstein_searcher import LevensteinSearcher
 
 
 logger = get_logger(__name__)
 
 
-@register('spelling_levenstein_searcher')
-class LeveSearcher(Component):
-    def __init__(self, words: Iterable[str], allow_spaces=True, max_distance=1, error_probability=1e-4, *args, **kwargs):
-        words = [word.strip().lower().replace('ё', 'е') for word in words]
+@register('spelling_levenstein')
+class LevensteinSearcherComponent(Component):
+    def __init__(self, words: Iterable[str], max_distance=1, error_probability=1e-4, *args, **kwargs):
+        words = list({word.strip().lower().replace('ё', 'е') for word in words})
         alphabet = sorted({letter for word in words for letter in word})
         self.max_distance = max_distance
         self.error_probability = log10(error_probability)
         self.vocab_penalty = self.error_probability * 2
-        self.searcher = LevensteinSearcher(alphabet, words, allow_spaces=allow_spaces, euristics=2)
+        self.searcher = LevensteinSearcher(alphabet, words, allow_spaces=True, euristics=2)
 
     def _infer_instance(self, tokens):
         candidates = []
